move prefetching stat names to StreamStatistics
